refactor: inline @automaker/policy-engine into server
Single consumer (authority-service.ts) didn't justify a standalone
package. Merged engine.ts + defaults.ts into
apps/server/src/services/policy-engine.ts and removed all workspace
references.
