+++ .changeset/config.json
@@ -11,7 +11,6 @@
       "@automaker/prompts",
       "@automaker/model-resolver",
       "@automaker/dependency-resolver",
-      "@automaker/policy-engine",
       "@automaker/spec-parser",
       "@automaker/flows",
       "@automaker/llm-providers",
