+++ apps/ui/scripts/prepare-server.mjs
@@ -27,7 +27,6 @@ const LOCAL_PACKAGES = [
   '@automaker/platform',
   '@automaker/model-resolver',
   '@automaker/dependency-resolver',
-  '@automaker/policy-engine',
   '@automaker/spec-parser',
   '@automaker/flows',
   '@automaker/llm-providers',
