--- libs/policy-engine/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "@automaker/policy-engine",
-  "version": "1.0.0",
-  "type": "module",
-  "description": "Policy engine for trust-based authorization in AutoMaker",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "scripts": {
-    "build": "tsc",
-    "watch": "tsc --watch",
-    "test": "vitest run",
-    "test:watch": "vitest"
-  },
-  "keywords": [
-    "automaker",
-    "policy",
-    "authorization",
-    "trust"
-  ],
-  "publishConfig": {
-    "access": "public"
-  },
-  "author": "AutoMaker Team",
-  "license": "SEE LICENSE IN LICENSE",
-  "engines": {
-    "node": ">=22.0.0 <23.0.0"
-  },
-  "dependencies": {
-    "@automaker/types": "1.0.0"
-  },
-  "devDependencies": {
-    "@types/node": "22.19.3",
-    "typescript": "5.9.3",
-    "vitest": "4.0.16"
-  }
-}
